Added dynamic darken and saturation for custom coloring of phases
